refactored example to use routes
